Fix: Sample code and normalize engine usage
The executor interface was accepting a session while the app_maker
was accepting an engine and building a sessions. Now the executor
accepts and engine instead of a session.

The sample server worked well enough, but it used an in memory
sqlite database. This made it incompatible with running an executor
agent. The database is now a sqlite file in the /tmp/ space.

Signed-off-by: Kevin Conway <kevinjacobconway@gmail.com>